status page now works. lets begin that pull request.
